[luci-interpreter] Remove type conversion warning (#4985)
This removes type conversion warning

ONE-DCO-1.0-Signed-off-by: Hyukjin Jeong <hj1.jeong@samsung.com>

+++ compiler/luci-interpreter/src/kernels/TestUtils.h
@@ -87,7 +87,7 @@ Tensor makeInputTensor(const Shape &shape, const std::vector<float> &scales,
   // inner dimensions that contains continuous data with one quantization type
   // outer dimensions that contains other dimensions
   size_t outer_dims_size = 1;
-  size_t quant_dim_size = static_cast<size_t>(shape.dim(quantized_dimension));
+  int32_t quant_dim_size = shape.dim(quantized_dimension);
   size_t inner_dims_size = 1;
   assert(quant_dim_size == scales.size());
   assert(quant_dim_size == zero_points.size());
@@ -102,7 +102,7 @@ Tensor makeInputTensor(const Shape &shape, const std::vector<float> &scales,
   std::vector<NativeT> quantized_data;
   quantized_data.reserve(shape.num_elements());
   for (size_t outer_it = 0; outer_it < outer_dims_size; ++outer_it)
-    for (size_t channel = 0; channel < quant_dim_size; ++channel)
+    for (int32_t channel = 0; channel < quant_dim_size; ++channel)
     {
       int32_t zero_point = zero_points[channel];
       float scale = scales[channel];
